% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_match.R
\name{add_group_id}
\alias{add_group_id}
\title{Match data or mc_nodes By Scenario or Groups}
\usage{
add_group_id(data_x, by, data_y = NULL)
}
\arguments{
\item{data_x}{Primary dataframe}

\item{by}{Vector of column names to group by}

\item{data_y}{Optional secondary dataframe for matching}
}
\value{
Dataframe or list of dataframes with added group IDs
}
\description{
Match data tables or Monte Carlo nodes and with different dimensions by scenario or groups
Add Group IDs to Data Frames
}
\details{
Creates a table summarizing origin farms with equivalent status
}
\examples{
add_group_id(purchase_data, by=c("animal_category", "pathogen", "mov_id", "farm_id"))
}
