% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_node_list.R
\name{add_prefix}
\alias{add_prefix}
\title{Add Prefix to Node Names}
\usage{
add_prefix(mcmodule, prefix = NULL, rewrite_module = NULL)
}
\arguments{
\item{mcmodule}{An mcmodule or a node_list object}

\item{prefix}{String to add as prefix of the new mcmodule mcnodes (preserves previous prefixes)}

\item{rewrite_module}{Name of a module to rewrite (new name equals prefix) - used after copying modules}
}
\value{
Returns either:
\itemize{
\item A node_list if input was a node_list
\item A model_object if input was an mcmodule
}
}
\description{
Adds a prefix to node_list names and all input nodes (non-prev)
}
\examples{
purchase <- add_prefix(
  mcmodule = transport_to,
  prefix = "transport_from",
  rewrite_module = "transport_to"
)
}
