% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_modules.R
\name{combine_modules}
\alias{combine_modules}
\title{Combine Two Modules}
\usage{
combine_modules(mcmodule_x, mcmodule_y)
}
\arguments{
\item{mcmodule_x}{First module to combine}

\item{mcmodule_y}{Second module to combine}
}
\value{
A combined mcmodule object containing elements from both input modules
}
\description{
This function combines two mcmodules into a single mcmodule, merging their data,
model expressions, node lists, and other components.
}
\examples{
combine_modules(mcmodule_x = purchase, mcmodule_y = fattening)

}
