% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmrisk-package.R
\docType{package}
\name{farmrisk-package}
\alias{farmrisk}
\alias{farmrisk-package}
\title{farmrisk: Farm Stochastic Risk Analysis}
\description{
Tools for performing quantitative risk analysis to prioritize farm external biosecurity measures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VetEpi-UAB/farmrisk}
  \item Report bugs at \url{https://github.com/VetEpi-UAB/farmrisk/issues}
}

}
\author{
\strong{Maintainer}: Natalia Ciria \email{natalia.ciria@uab.cat} (\href{https://orcid.org/0009-0006-2669-6634}{ORCID})

Other contributors:
\itemize{
  \item Alberto Allepuz \email{alberto.allepuz@uab.cat} (\href{https://orcid.org/0000-0003-3518-1991}{ORCID}) [thesis advisor]
  \item Giovanna Ciaravino \email{giovanna.ciaravino@uab.cat} (\href{https://orcid.org/0000-0002-5796-8093}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
