% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_totals.R
\name{get_agg_totals}
\alias{get_agg_totals}
\title{Calculate Aggregated Totals Across Groups}
\usage{
get_agg_totals(
  mcmodule,
  mcnode,
  keys_names = c("pathogen", "farm_id", "scenario_id"),
  suffix = "agg",
  data_name = NULL,
  summary = calc_summary,
  agg_variates = FALSE
)
}
\arguments{
\item{mcmodule}{mcmodule object containing the data}

\item{mcnode}{name of node to aggregate}

\item{keys_names}{grouping variables for aggregation}

\item{suffix}{suffix for output node name}

\item{data_name}{name/index of data to use}

\item{summary}{logical, whether to calculate summaries}

\item{agg_variates}{logical, whether to aggregate variates}
}
\value{
Updated mcmodule with new aggregated nodes
}
\description{
Aggregates node values across specified grouping variables
}
\examples{
get_agg_totals(mcmodule = purchase_origin,
               mcnode = "a_inf_all")
}
