% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_custom_table.R
\name{get_columns}
\alias{get_columns}
\title{Column selection keeping ids}
\usage{
get_columns(data, col, pathway = NULL, module = NULL)
}
\arguments{
\item{data}{Dataframe}

\item{col}{Columns to extract}

\item{pathway}{Optional pathway filter}

\item{module}{Optional module filter}
}
\value{
Filtered dataframe with selected columns
}
\description{
Extracts specific columns from datasets while preserving id fields.
}
\examples{
get_columns(data = mov, col = "purchase_single_date")
}
