% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mc_inputs.R
\name{get_mc_inputs}
\alias{get_mc_inputs}
\title{Get Monte Carlo Input Variables from CSV Files}
\usage{
get_mc_inputs(path = input_path, mctable = mcnode_admin)
}
\arguments{
\item{path}{Character string specifying the path to input files. Default is "input_files/".}

\item{mctable}{Data frame containing the list of Monte Carlo objects and their definitions.
Default is mcnode_admin.}
}
\value{
Character vector of column names that are MC node inputs.
}
\description{
This function searches through CSV files in a specified directory to find variables
that are inputs for Monte Carlo nodes (mcnodes) defined in a Monte Carlo table.
}
\examples{
\dontrun{
  mc_inputs <- get_mc_inputs()
  mc_inputs <- get_mc_inputs(path = "my_files/", mctable = my_mc_table)
}
}
