% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_model.R
\name{get_node_list}
\alias{get_node_list}
\title{Match Data or Monte Carlo Nodes By Scenario or Groups}
\usage{
get_node_list(
  model_expression,
  param_names = NULL,
  mctable = mcnode_admin,
  path = "input_files/"
)
}
\arguments{
\item{model_expression}{An R expression containing model calculations}

\item{param_names}{Optional named vector for parameter renaming}

\item{mctable}{Reference table for MC nodes, defaults to mcnode_admin}

\item{path}{Path to input files, defaults to "input_files/"}
}
\value{
A list of class "mcnode_list" containing node information
}
\description{
Functions for managing Monte Carlo nodes in a modular simulation framework.
}
\details{
Create Node List from Model Expression

This function creates a list of nodes based on a given model expression, handling input,
output, and previous nodes with their respective properties and relationships.
}
\examples{
get_node_list(model_expression = indir_contact_expression)
}
