% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_code.R
\name{get_region_level}
\alias{get_region_level}
\title{Get region level of region_code}
\usage{
get_region_level(data, key = "region_code")
}
\arguments{
\item{data}{dataframe or vector containing region codes}

\item{key}{column name containing region codes if data is a dataframe}
}
\value{
Region code GID level
}
\description{
Get region level of region_code
}
\examples{
get_region_level(value = "ESP.2_1")
}
