% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_code.R
\name{gid_levels}
\alias{gid_levels}
\title{Get all region levels of region_code}
\usage{
gid_levels(data, key = "region_code", only_higher = FALSE)
}
\arguments{
\item{data}{dataframe or vector with region codes in GID format}

\item{key}{column name containing region codes if data is a dataframe}

\item{only_higher}{logical, whether to return only higher level codes}
}
\value{
List of region code GID levels
}
\description{
Get all region levels of region_code
}
\examples{
get_all_region_levels(value = "ESP.2_1")
}
