% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_match.R
\name{group_match}
\alias{group_match}
\title{Match Groups Between Datasets}
\usage{
group_match(x, y, by = NULL)
}
\arguments{
\item{x}{First dataset}

\item{y}{Second dataset}

\item{by}{Optional grouping variables}
}
\value{
Matched dataset
}
\description{
Match Groups Between Datasets
}
\examples{
group_match(dataset1, dataset2, by=c("category", "group"))
}
