% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_code.R
\name{hjoin_region}
\alias{hjoin_region}
\title{Hierarchical Join by region_code}
\usage{
hjoin_region(x, y, by = NULL, keys = NULL, add_keys = NULL)
}
\arguments{
\item{x, y}{A pair of data frames to merge with region GID codes}

\item{by}{Column names to join by (NULL for natural join)}

\item{keys}{Additional keys for grouping}

\item{add_keys}{Extra keys to include}
}
\value{
Merged dataframe
}
\description{
Based on: Barks P (2022). hmatch: Tools for Cleaning and Matching
Hierarchically-Structured Data.
}
\examples{
hjoin_region(x = visit_data, y = visit_region)
}
