% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_summary.R
\name{input_keys}
\alias{input_keys}
\title{Monte Carlo Node Keys and Summaries}
\usage{
input_keys(x)
}
\arguments{
\item{x}{mcnode or character string}
}
\value{
character vector of key names
}
\description{
Collection of functions for handling Monte Carlo Nodes using their metadata found in their mcmodule
Get mcnode admin inputs key names
}
\examples{
input_keys(mcnode = inf_dc)
}
