% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_match.R
\name{mc_agg_match}
\alias{mc_agg_match}
\title{Match Monte Carlo Nodes with Null Values}
\usage{
mc_agg_match(mcmodule, mc_name_x, mc_name_y, keys_names = c("pathogen"))
}
\arguments{
\item{mcmodule}{Monte Carlo module}

\item{mc_name_x}{First node name}

\item{mc_name_y}{Second node name}

\item{keys_names}{Names of key columns}
}
\value{
List containing matched nodes and index
}
\description{
Matches two mc_nodes with differing dimensions, handling null values
}
\examples{
mc_agg_match(mcmodule = intro, 
             mc_name_x="no_purchase_inf_agg", 
             mc_name_y="b_entry_agg")
}
