% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_match.R
\name{mc_match}
\alias{mc_match}
\title{Match Monte Carlo Nodes}
\usage{
mc_match(mcmodule, mc_name, data)
}
\arguments{
\item{mcmodule}{Monte Carlo module}

\item{mc_name}{Name of the Monte Carlo node}

\item{data}{Input data}
}
\value{
Modified Monte Carlo node
}
\description{
Matches monte carlo nodes with differing dimensions assuming the new variates
are equal for scenario 0 for the smallest node
}
\examples{
mc_match(mcmodule = purchase, mc_name="fattening_b_indir_contact_all", data=quarantine_data)
}
