% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_summary.R
\name{mc_summary}
\alias{mc_summary}
\title{Get mcnode summary statistics}
\usage{
mc_summary(
  data,
  mcnode,
  sep_keys = TRUE,
  na_value = NA,
  keys_names = NULL,
  mc_name = NULL,
  by_keys = FALSE
)
}
\arguments{
\item{data}{data frame containing mnode inputs}

\item{mcnode}{mcnode object}

\item{sep_keys}{logical, separate keys in different columns}

\item{na_value}{value to exclude from summary}

\item{keys_names}{vector of key names for grouping}

\item{mc_name}{character, name of mcnode}

\item{by_keys}{logical, whether to summarize by keys}
}
\value{
summary data frame
}
\description{
Get mcnode summary statistics
}
\examples{
mc_summary(mcnode = h_prev)
}
