% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_wif.R
\name{output_wif}
\alias{output_wif}
\title{Process what-if scenario output}
\usage{
output_wif(data, scenario, data_wif, new_wif, wif_only)
}
\arguments{
\item{data}{Original data frame}

\item{scenario}{Scenario name}

\item{data_wif}{Modified data for new scenario}

\item{new_wif}{New values to apply}

\item{wif_only}{Whether to return only new scenario}
}
\value{
Processed data frame with scenario changes applied
}
\description{
Process what-if scenario output
}
