% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_input_files.R
\name{process_json_files}
\alias{process_json_files}
\title{Process JSON files to CSV}
\usage{
process_json_files(user_path, envir)
}
\arguments{
\item{user_path}{Character string. Path to user directory}

\item{envir}{Environment. envir environment to store data}
}
\description{
Process JSON files to CSV
}
