% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_simulation.R
\name{save_mc_summary}
\alias{save_mc_summary}
\title{Save Monte Carlo Module Summary}
\usage{
save_mc_summary(mcmodule, mcnodes)
}
\arguments{
\item{mcmodule}{List containing Monte Carlo module data}

\item{mcnodes}{Character vector of node names to process}
}
\value{
None (saves files to disk)
}
\description{
Saves summary tables from Monte Carlo module nodes to CSV files
}
