% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_simulation.R
\name{save_table_summary}
\alias{save_table_summary}
\title{Save Table Summary}
\usage{
save_table_summary(tables)
}
\arguments{
\item{tables}{Named list of data frames to save}
}
\value{
None (saves files to disk)
}
\description{
Saves data tables to CSV files
}
