% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_wif.R
\name{set_up_wif}
\alias{set_up_wif}
\title{Initialize data for what-if scenarios}
\usage{
set_up_wif(data)
}
\arguments{
\item{data}{Data frame to initialize}
}
\value{
Initialized data frame with homogeneous group IDs and wif flag
}
\description{
Initialize data for what-if scenarios
}
