% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_custom_table.R
\name{tidy_animal_table}
\alias{tidy_animal_table}
\title{Tidy animal table}
\usage{
tidy_animal_table(data = mov, pathway, module = NULL)
}
\arguments{
\item{data}{Dataframe containing animal movements records (bsg by default)}

\item{pathway}{Pathway filter}

\item{module}{Optional module filter}
}
\value{
Dataframe with animal counts by category
}
\description{
Restructures animal movements data into a table
}
\examples{
tidy_animals_table(data=mov,pathway="purchase",module="quarantine")
}
