% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_custom_table.R
\name{tidy_census_table}
\alias{tidy_census_table}
\title{Tidy animal census}
\usage{
tidy_census_table(data = bsg, census_key = "farm_census")
}
\arguments{
\item{data}{Dataframe containing animal census records (bsg by default)}

\item{census_key}{Identifier for census data ("farm_census" by default)}
}
\value{
Dataframe with animal counts by category
}
\description{
Restructures animal census data into a table
}
\examples{
tidy_census_table(data = bsg, census_key = "farm_census")
}
