% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_custom_table.R
\name{tidy_group}
\alias{tidy_group}
\title{Tidy farm groups}
\usage{
tidy_group(data)
}
\arguments{
\item{data}{Dataframe containing one rows for each individual farm}
}
\value{
Summarized with the number of farms and the number of animals in each homogeneous group
}
\description{
Aggregates farms with the same characteristics to create homogeneous groups
}
\examples{
tidy_group(data)
}
