% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_custom_table.R
\name{tidy_panel}
\alias{tidy_panel}
\title{Tidy panel}
\usage{
tidy_panel(data = bsg, panel)
}
\arguments{
\item{data}{Dataframe containing a panel (bsg by default)}

\item{panel}{Name of the panel}
}
\value{
Dataframe with rows by panel entries
}
\description{
Creates a long dataframe from a wide dataframe with series of questions repeated n times.
}
\examples{
tidy_from_panel(data=bsg,panel="veh")
}
