% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_custom_table.R
\name{tidy_prefix}
\alias{tidy_prefix}
\title{Tidy and Merge User and Admin Inputs}
\usage{
tidy_prefix(data, module = NULL, pathway = NULL, rm_prefix = TRUE)
}
\arguments{
\item{data}{Dataframe}

\item{module}{Optional module filter}

\item{pathway}{Optional pathway filter}

\item{rm_prefix}{Logical indicating whether to remove prefix from column names (TRUE by default)}
}
\value{
A cleaned dataframe with standardized column names
}
\description{
A collection of functions for handling bsg and mov inputs and
merge them with admin parameters by pathways and modules. This functions are based on survey.js
answer structure and and its conversion to csv.

Tidy prefix

Processes dataframe columns by selecting those with a specific prefix and optionally removing that prefix.
}
\examples{
tidy_prefix(data = bsg, module = "quarantine")
}
