% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_custom_table.R
\name{tidy_status_table}
\alias{tidy_status_table}
\title{Tidy health status data}
\usage{
tidy_status_table(data = mov, pathway, module)
}
\arguments{
\item{data}{Dataframe containing Tidy health status data records (mov by default)}

\item{pathway}{Pathway filter}

\item{module}{Module filter}
}
\value{
Dataframe with health status by movement, pathogen and farm
}
\description{
Restructures Tidy health status data into a table
}
\examples{
tidy_status_table(data = mov, pathway = "purchase", module = "quarantine")
}
