% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_custom_table.R
\name{tidy_test_table}
\alias{tidy_test_table}
\title{Tidy diagnostic test data}
\usage{
tidy_test_table(data = mov, pathway, module)
}
\arguments{
\item{data}{Dataframe containing diagnostic test records (mov by default)}

\item{pathway}{Pathway filter}

\item{module}{Module filter}
}
\value{
Dataframe with diagnostic tests by movement, pathogen and farm
}
\description{
Restructures diagnostic test data into a table
}
\examples{
tidy_test_table(data = mov, pathway = "purchase", module = "origin")
}
