% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_wif.R
\name{wif_clean}
\alias{wif_clean}
\title{Cleaning and Disinfection Scenario}
\usage{
wif_clean(data, scenario = NULL, wif_only = FALSE, prev_wif = NULL)
}
\description{
Simulates implementation of cleaning and disinfection protocols
}
