% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_wif.R
\name{wif_exclusive_material}
\alias{wif_exclusive_material}
\title{Exclusive Material Usage Scenario}
\usage{
wif_exclusive_material(
  data,
  scenario = NULL,
  wif_only = FALSE,
  prev_wif = NULL
)
}
\description{
Simulates use of exclusive equipment and materials
}
