% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_wif.R
\name{wif_high_waterer}
\alias{wif_high_waterer}
\title{High Waterers Scenario}
\usage{
wif_high_waterer(data, scenario = NULL, wif_only = FALSE, prev_wif = NULL)
}
\arguments{
\item{data}{Data frame containing the original dataset}

\item{scenario}{Character string specifying the scenario name (optional)}

\item{wif_only}{Logical indicating if only what-if records should be returned}

\item{prev_wif}{Character string specifying a previous what-if scenario (optional)}
}
\value{
Modified data frame with what-if scenario results
}
\description{
Simulates elevation of water points to prevent disease transmission
}
