% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_wif.R
\name{wif_no_dc}
\alias{wif_no_dc}
\title{No Direct Contact Scenario}
\usage{
wif_no_dc(data, scenario = NULL, wif_only = FALSE, prev_wif = NULL)
}
\description{
Simulates prevention of direct animal contact
}
