% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_wif.R
\name{wif_no_fomites}
\alias{wif_no_fomites}
\title{No Fomites Scenario}
\usage{
wif_no_fomites(
  data,
  fomites = NULL,
  scenario = NULL,
  wif_only = FALSE,
  prev_wif = NULL
)
}
\arguments{
\item{fomites}{Character vector specifying types of fomites to control}
}
\description{
Simulates prevention of fomite introduction
}
