% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_wif.R
\name{wif_no_share_pasture}
\alias{wif_no_share_pasture}
\title{No Shared Pasture Scenario}
\usage{
wif_no_share_pasture(data, scenario = NULL, wif_only = FALSE, prev_wif = NULL)
}
\description{
Simulates prevention of shared pasture usage
}
