% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_wif.R
\name{wif_no_share_veh}
\alias{wif_no_share_veh}
\title{No Shared Vehicle Scenario}
\usage{
wif_no_share_veh(data, scenario = NULL, wif_only = FALSE, prev_wif = NULL)
}
\description{
Simulates avoiding shared transport with other farms
}
