% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_wif.R
\name{wif_own_veh}
\alias{wif_own_veh}
\title{Own Vehicle Usage Scenario}
\usage{
wif_own_veh(data, scenario = NULL, wif_only = FALSE, prev_wif = NULL)
}
\description{
Simulates exclusive use of own vehicles
}
