% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_wif.R
\name{wif_quarantine}
\alias{wif_quarantine}
\title{Quarantine Scenario}
\usage{
wif_quarantine(data, scenario = NULL, wif_only = FALSE, prev_wif = NULL)
}
\description{
Simulates implementation of quarantine protocols for new animals
}
